Pin third-party GitHub Actions (#52)
GitHub
[recommends](https://docs.github.com/en/actions/reference/security/secure-use#using-third-party-actions)
pinning third-party Actions to commit SHAs instead of versions, which
are just Git tags/branches and can introduce a supply-chain attack.

I didn't pin the Ruby one, since we assume it to be "trusted" (and want
to keep the version up to date), but if you think we should pin it as
well, will do.